+++ src/util/index.js
@@ -1,2 +1,3 @@
 export * from './show';
 export * from './rating';
+export * as schedule from './schedule';